Fix many errors that occur on 32-bit systems under UBSan (#804)
This commit fixes fixes a large number of runtime errors that occur
when ksh is compiled with '-fsanitize=undefined' on Linux 32-bit
using GCC.

Noted changes:
- Switched out long and double for Sflong_t and Sfdouble_t in many
  places to fix runtime errors under UBSan. (In certain places
  (u)int64_t was used to specify exactly 64-bits rather than
  intmax_t.) The following runtime errors should no longer occur
  (line numbers/positions as of commit e67af956):

  * left shift of 1 by 31 places cannot be represented in type
    'long int': macro.c:2489:9, args.c:252:4, args.c:272:4,
    args.c:175:60, args.c:607:10, args.c:622:38
  * left shift of 12 by 28 places cannot be represented in type
    'long int': pty.c:355:12
  * signed integer overflow: 1734423994 - -1213168416 cannot be
    represented in type 'long int': xec.c:2359:4

- For mamake, use 'unsigned long long' instead of 'uint64_t' to
  avoid requiring stdint.h.

- Remove !timeofday code. All modern systems have gettimeofday,
  which makes this code unused and an unnecessary maintenance
  burden.

- Define _TIME_BITS=64 to fix integer overflows in the time keyword
  and likely elsewhere (ksh assumes time_t is 64-bits).

- Moved the AST_LC_* bitflags into the 32-bit range.

- Don't assume 'long long' and 'intmax_t' are the same. While
  intmax_t is in practice the same as 'long long', this is not
  guaranteed by the C standard (it's permitted for these types to
  differ). For correctness, only use %jd and %ju with Sflong_t (aka
  intmax_t), and enforce this distinction in SFIO's handling of %ll
  and %j. (The extra if statements are optimized away on platforms
  where the types are identical.)

- Use suseconds_t for tv->usec and enforce 64-bit time_t in the
  libast feature tests.

- Casted more PIDs to Sflong_t when using PIDs with format
  specifiers.

- Use the %zu format specifier with size_t vars rather than
  unnecessary casts.

+++ src/cmd/ksh93/include/defs.h
@@ -167,17 +167,17 @@ extern char		*sh_getcwd(void);
 #endif
 #define sh_translate(s)	_sh_translate(ERROR_dictionary(s))
 
-#define WBITS		(sizeof(long)*8)
+#define WBITS		(sizeof(uint64_t)*8)
 #define WMASK		(0xff)
 
 #if SHOPT_SCRIPTONLY
-#define is_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] & (1L << ((x) % WBITS))) )
-#define on_option(s,x)	( (x)==SH_INTERACTIVE || (x)==SH_HISTORY ? errormsg(SH_DICT,ERROR_exit(1),e_scriptonly) : ((s)->v[((x)&WMASK)/WBITS] |= (1L << ((x) % WBITS))) )
-#define off_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] &= ~(1L << ((x) % WBITS))) )
+#define is_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] & ((uint64_t)1 << ((x) % WBITS))) )
+#define on_option(s,x)	( (x)==SH_INTERACTIVE || (x)==SH_HISTORY ? errormsg(SH_DICT,ERROR_exit(1),e_scriptonly) : ((s)->v[((x)&WMASK)/WBITS] |= ((uint64_t)1 << ((x) % WBITS))) )
+#define off_option(s,x)	((x)==SH_INTERACTIVE || (x)==SH_HISTORY ? 0 : ((s)->v[((x)&WMASK)/WBITS] &= ~((uint64_t)1 << ((x) % WBITS))) )
 #else
-#define is_option(s,x)	((s)->v[((x)&WMASK)/WBITS] & (1L << ((x) % WBITS)))
-#define on_option(s,x)	((s)->v[((x)&WMASK)/WBITS] |= (1L << ((x) % WBITS)))
-#define off_option(s,x)	((s)->v[((x)&WMASK)/WBITS] &= ~(1L << ((x) % WBITS)))
+#define is_option(s,x)	((s)->v[((x)&WMASK)/WBITS] & ((uint64_t)1 << ((x) % WBITS)))
+#define on_option(s,x)	((s)->v[((x)&WMASK)/WBITS] |= ((uint64_t)1 << ((x) % WBITS)))
+#define off_option(s,x)	((s)->v[((x)&WMASK)/WBITS] &= ~((uint64_t)1 << ((x) % WBITS)))
 #endif /* SHOPT_SCRIPTONLY */
 #define sh_isoption(x)	is_option(&sh.options,x)
 #define sh_onoption(x)	on_option(&sh.options,x)
